requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.24.3
sentence-transformers==2.2.2
aiohttp==3.8.5
faiss-cpu==1.7.3
pymongo==4.4.1
python-dotenv==1.0.0
